Infrastructure updates
This commit updates the infrastructure to use the latest version of the
Terraform provider for AWS. This is a breaking change, as the provider
has been updated to version 3.0.0, which is not backwards-compatible with
the previous version.

The following changes were made:

- The provider.tf file was updated to use the new version of the provider.
- The lambda_function.py file was updated to use the new version of the
  provider.
- The outputs.tf file for the Prometheus and SonarQube modules was updated
  to use the new version of the provider.
- The topic.tf file was updated to use the new version of the provider.

Please note that this change requires a manual update to the state file,
as the new version of the provider is not backwards-compatible with the
previous version. This can be done by running the following command:

    terraform state replace-provider registry.terraform.io/-/aws registry.terraform.io/hashicorp/aws

Once the state file has been updated, the infrastructure can be updated
using the following command:

    terraform apply

Please let me know if you have any questions or concerns about this change.

Signed-off-by: John Doe <